+++ src/app/researcher/study/request/[memberIdentifier]/study-proposal.tsx
@@ -1,126 +1,71 @@
 'use client'
 
-import React, { FC } from 'react'
-import { useUser } from '@clerk/nextjs'
-import { Anchor, Divider, FileInput, Group, Paper, Stack, Text, TextInput, Title, useMantineTheme } from '@mantine/core'
-import { FileDoc, FilePdf, FileText, UploadSimple } from '@phosphor-icons/react/dist/ssr'
-import { UseFormReturnType } from '@mantine/form'
-import { StudyProposalFormValues } from '@/app/researcher/study/request/[memberIdentifier]/study-proposal-schema'
+import React from 'react'
+import { Button, Group, Stack } from '@mantine/core'
+import { notifications } from '@mantine/notifications'
+import { CancelButton } from '@/components/cancel-button'
+import { useForm } from '@mantine/form'
+import {
+    StudyProposalFormValues,
+    studyProposalSchema,
+    zodResolver,
+} from '@/app/researcher/study/request/[memberIdentifier]/study-proposal-schema'
+import { StudyProposalForm } from '@/app/researcher/study/request/[memberIdentifier]/study-proposal-form'
+import { UploadStudyJobCode } from '@/app/researcher/study/request/[memberIdentifier]/upload-study-job-code'
+import { useMutation } from '@tanstack/react-query'
+import { onCreateStudyAction } from '@/app/researcher/study/request/[memberIdentifier]/actions'
+import { useRouter } from 'next/navigation'
 
-export const StudyProposalForm: FC<{
-    studyProposalForm: UseFormReturnType<StudyProposalFormValues>
-}> = ({ studyProposalForm }) => {
-    const theme = useMantineTheme()
-    const color = theme.colors.purple[6]
+export const StudyProposal: React.FC<{ memberId: string }> = ({ memberId }) => {
+    const router = useRouter()
 
-    const getFileUploadIcon = (color: string, fileName?: string | null) => {
-        if (!fileName) return <UploadSimple size={14} color={color} weight="fill" />
-        const Icons: [RegExp, React.ReactNode][] = [
-            [/\.docx?$/i, <FileDoc key="doc" size={14} color={color} />],
-            [/\.txt$/i, <FileText key="txt" size={14} color={color} />],
-            [/\.pdf$/i, <FilePdf key="pdf" size={14} color={color} />],
-        ]
-        const matchedIcon = Icons.find(([re]) => re.test(fileName))?.[1]
-        return matchedIcon || <UploadSimple size={14} color={color} weight="fill" />
-    }
+    const { mutate: createStudy } = useMutation({
+        mutationFn: async (formValues: StudyProposalFormValues) => {
+            return await onCreateStudyAction({ memberId, studyInfo: formValues })
+        },
+        onSuccess() {
+            notifications.show({
+                title: 'Study Proposal Submitted',
+                message:
+                    'Your proposal has been successfully submitted to the reviewing organization. Check your dashboard for status updates.',
+                color: 'green',
+            })
+            router.push(`/researcher/dashboard`)
+        },
+        onError(error) {
+            // TODO server action max filesize error doesn't propagate through here...
+            console.error(error)
+            notifications.show({ message: String(error), color: 'red' })
+        },
+    })
 
-    const fileUpload = getFileUploadIcon(color, studyProposalForm.values.descriptionDocument?.name ?? '')
-    const irbFileUpload = getFileUploadIcon(color, studyProposalForm.values.irbDocument?.name ?? '')
-    const agreementFileUpload = getFileUploadIcon(color, studyProposalForm.values.agreementDocument?.name ?? '')
-    const { user } = useUser()
+    const studyProposalForm = useForm<StudyProposalFormValues>({
+        validate: zodResolver(studyProposalSchema),
+        validateInputOnBlur: true,
+        initialValues: {
+            title: '',
+            piName: '',
+            irbDocument: null,
+            descriptionDocument: null,
+            agreementDocument: null,
+            codeFiles: [],
+        },
+    })
 
+    // TODO Can we make a stepper https://mantine.dev/core/stepper/
     return (
-        <Paper p="md">
-            <Title order={4} c="purple">
-                Study Proposal
-            </Title>
-            <Divider my="sm" mt="sm" mb="md" />
-            <Text>
-                This section is key to your proposal, as it defines the analysis that will generate the results you’re
-                intending to obtain from the Member’s data. Upload any necessary files to support your analysis. In this
-                iteration, we currently support .r and .rmd files.
-            </Text>
-            <Stack gap="lg" mt="md">
-                {/* TODO flesh out with UX/do when hifi-s ready */}
-                <Group gap="xl">
-                    <Text>Study Title</Text>
-                    <TextInput
-                        aria-label="Study Title"
-                        placeholder="Enter a title (max. 50 characters)"
-                        {...studyProposalForm.getInputProps('title')}
-                    />
-                </Group>
-
-                <Group gap="xl">
-                    <Text>Study Lead</Text>
-                    <TextInput aria-label="Study Lead" disabled value={user?.fullName ?? ''} />
-                </Group>
-
-                <Group gap="xl">
-                    <Text>Principal Investigator</Text>
-                    <TextInput aria-label="Principal Investigator" {...studyProposalForm.getInputProps('piName')} />
-                </Group>
-
-                <Group>
-                    <Text>Study Description</Text>
-                    {fileUpload}
-                    <Stack gap={0}>
-                        <FileInput
-                            name="descriptionDocument"
-                            component={Anchor}
-                            aria-label="Upload Study Description Document"
-                            placeholder="Upload Document"
-                            clearable
-                            accept=".doc,.docx,.txt,.pdf"
-                            {...studyProposalForm.getInputProps('descriptionDocument')}
-                        />
-                        <Text size="xs" c="dimmed">
-                            Accepted formats: doc, docx, pdf
-                        </Text>
-                    </Stack>
-                </Group>
-
-                <Group gap="xs">
-                    <Text>IRB Document</Text>
-                    {irbFileUpload}
-                    <Stack gap={0}>
-                        <FileInput
-                            name="irbDocument"
-                            color="purple"
-                            component={Anchor}
-                            aria-label="Upload IRB Document"
-                            placeholder="Upload Document"
-                            clearable
-                            accept=".doc,.docx,.txt,.pdf"
-                            key={studyProposalForm.key('irbDocument')}
-                            {...studyProposalForm.getInputProps('irbDocument')}
-                        />
-                        <Text size="xs" c="dimmed">
-                            Accepted formats: doc, docx, pdf
-                        </Text>
-                    </Stack>
-                </Group>
-
-                <Group gap="xs">
-                    <Text>Agreement Document</Text>
-                    {agreementFileUpload}
-                    <Stack gap={0}>
-                        <FileInput
-                            name="agreementDocument"
-                            aria-label="Upload Agreement Document"
-                            component={Anchor}
-                            placeholder="Upload Document"
-                            clearable
-                            accept=".doc,.docx,.txt,.pdf"
-                            key={studyProposalForm.key('agreementDocument')}
-                            {...studyProposalForm.getInputProps('agreementDocument')}
-                        />
-                        <Text size="xs" c="dimmed">
-                            Accepted formats: doc, docx, pdf
-                        </Text>
-                    </Stack>
+        <form onSubmit={studyProposalForm.onSubmit((values: StudyProposalFormValues) => createStudy(values))}>
+            <Stack>
+                <StudyProposalForm studyProposalForm={studyProposalForm} />
+                <UploadStudyJobCode studyProposalForm={studyProposalForm} />
+                <Group gap="xl" justify="flex-end">
+                    {/* TODO Talk about removing cancel button, next/back buttons, submit button layout with UX */}
+                    <CancelButton isDirty={studyProposalForm.isDirty()} />
+                    <Button disabled={!studyProposalForm.isValid} type="submit" variant="filled">
+                        Submit
+                    </Button>
                 </Group>
             </Stack>
-        </Paper>
+        </form>
     )
 }
